+++ lib/chart/render/decorations/vertical_axis_decoration.dart
@@ -1,6 +1,6 @@
 part of charts_painter;
 
-typedef AxisValueFromIndex = String Function(int index);
+typedef AxisValueFromIndex = String Function(double index);
 
 /// Position of legend in [VerticalAxisDecoration]
 enum VerticalLegendPosition {
@@ -143,7 +143,7 @@ class VerticalAxisDecoration extends DecorationPainter {
       String? _text;
 
       try {
-        _text = valueFromIndex((axisStep * i).toInt());
+        _text = valueFromIndex((axisStep * i));
       } catch (e) {
         /// Invalid value, index out of range can happen here.
       }
